requirements.txt
flask==3.1.0
google-generativeai==0.8.5
gunicorn==23.0.0
